Refactor some parts of loggedOnly
- Rename the middlewares directory to middleware
- Fix bad import statement in routes/users/index.ts
- Add some comments in loggedOnly
- Catch JsonWebToken's throw to produce a response with a correct
http code
