+++ tsconfig.json
@@ -28,7 +28,7 @@
         "sources/api/views/UserView.ts",
 
         // Middleware
-        "sources/api/middlewares/loggedOnly.ts",
+      "sources/api/middleware/loggedOnly.ts",
 
         // Utils
         "sources/utils/db.ts",
